+++ CHANGELOG.md
@@ -1,14 +1,22 @@
+2.1.0, 2019-03-15:
+- Enable signature algorithm selection (SHA-1 or HMAC-SHA-256).
+- Added Spanish translation.
+- Improve plugin translations.
+- [prodfaq] Fix notice about shifting the shop to production mode.
+- Display card brand user choice if any in backend order details.
+- [technical] Manage enabled/disabled features by plugin variant.
+
 2.0.0, 2017-06-05:
-- Adding logs to file.
 - Bug fix: amount sent to the plaform was not the same as one shown in merchant website (rounding error).
 - Single payment and payment in installments in the same module.
 - Possibility to configure several options for payment in installments.
 - Rename module folders and files.
 - Adding german translations.
 - Remove control over certificate format modified on the platform.
+- Adding logs to file.
 
 1.1.0, 2015-12-03:
 - Compatibility with commercial version of Hikashop.
 
 1.0, 2013-08-20:
-- New PayZen payment module for HikaShop_2.2.0
+- New PayZen payment module for HikaShop 2.2.0